Added NumberStabilisation.vhd and assoc. software
Added the number stabilisation module and associated control software in
MATLAB.  The new module calculates the number of microwave pulses to
create after every dispersive probing pulse via a difference between
a user-supplied maximum number of pulses, which should remove ALL atoms,
and a computed value which is the division of a 'target' value with the
measured value.

There should also be manual control of the microwave pulses, and these
can be triggered by software or hardware.  I have added a hardware
trigger to the system for the dispersive probing and microwave pulse
generators.

Changed the fields for the t_control VHDL type to have start, stop, and
enable fields.  This should make it more useful?

Changed behaviour of MATLAB DPFeedbackParameter class so that it can
handle parameters that have been split across registers.  This was needed
because the so-called quadTarget parameter is the product of the actual
target value and the bare number of pulses, and thus is 40 bits wide.
This was too long for a single register, so it had to be split across two.
New code should behave sensibly.