Created VHDL code for signal processing via peak detection
New method for calculating the signal from DP pulses.  Idea is
to demodulate the RF signal down to a few MHz and then determine
the amplitude of the signal by finding the maximum and minimum
values in a given window.  As long as the window is longer than
one period, a maximum and minimum will be present.  This gives the
peak-to-peak amplitude.  This amplitude can then be used for down-
stream processing.

This very much needs to be tested!