Added dpOnWhenShutterClosed signal to topmod
Also fixed issue with pulse saving getting out of sync, which led to 1
more pulse being saved than should have been.  This was caused by the
reset signal being driven as synchronous with sysClk but needing to be
synchronous with adcClk.  Fixed this by creating a rising-edge detector
for the reset signal in SaveADCData.vhd