Major change to allow for auxiliary measurements
THIS VERSION IS NOT READY FOR USE

Determined that I need to be able to independently measure the gain of
the avalanche photo-detectors that I use for acquiring Faraday rotation
data.  To do so requires an auxiliary measurement.  I've changed the
structure so that the system will acquire data from interleaved pulses
where one pulse is the actual signal and the second is the auxiliary
measurement for determining the APD gains.

In addition, I've changed how parameters are sent to some modules.
Rather than individual parameters, some of them are now arrays.  The
way it is currently written may not be syntactically correct, so further
edits will be required.