Major revision to include on-the-fly gains
Added VHDL code and IP sources to allow for two different chains of
signal acquisition chains: one for the actual signal and an auxiliary
path that allows for the gains of the two detectors to be computed.

There are now four memory elements rather than 2: 2 for the signal chain
and 2 for the auxiliary chain.  These are still accessed in the normal
way.

For modules where there are multiple 32-bit parameters, these parameters
are now passed using the t_param_reg_array(natural range <>) type.

Building blocks of this code have been tested in simulation but not in
practice.  This code has been compiled into a working design.