Continued work on software and VHDL - in progress
Continued work on MATLAB software with a number of adjustments to the
python socket server and also the C code for fetching data from the
FPGA.  Had a lot of trouble with FPGA not registering all the pulses,
and I resolved this eventually by changing the clock for the PulseGen
module instances from sysClk to adcClk.  There's some subtlety to the
timing that I just don't get.

The signal and auxiliary data are now stored in the DPFeedback class
using the DPFeedbackData class just so that I don't have to have a bunch
of doubled properties just for this purpose.

I also added a shell function that gets the IP address that starts with
'172.22' on the Red Pitaya.  Previously I just used 'hostname -I' but
that started to report two IP addresses, so now there is a shell function
for that purpose.