Major changes to finish up dual acquisition
Dual acquisition seems to work now.  Signal value, computed as the ratio
of the difference in the two signals to their sum, is now saved in a
block memory (13 bits wide).  This ratio is computed using the full
precision of all the relevant signals, as doing so with less seemed to
cause issues when comparing to values calculated in MATLAB.

I had to increase the multiplication and division latencies in the
module ComputeSignal because if they are shorter then the returned value
is incorrect (actually it looks like it's delayed by one sample).