Major changes so power meas works with new system
Power measurement now uses the DualChannelAcquisition module to acquire
two sets of dual ADC data just like the feedback system.  Lots of issues
with the power measurement system not recording the right number of pulses
AND with it not starting data acquisition at the right time.  The latter
issue seemed to be due to the auto reset feature, and that was fixed
(somehow) by changing the autoResetCount + 1 statement to
autoResetCount + AUTO_RESET_INCR where both variables/constants are
unsigned types of the same length.  It seemed like the problem was due
to an issue with the "+" operator for unsigned and integer types.

The former issue, with missed pulses, was fixed by changing where the
delayCount signal was reset to its initial value of 1.  As far as I can
tell, this should not have made a difference, but it did so I'm going
with it.