Changed how ratio is calculated
New scheme for calculating the Faraday signal ratio which should be
independent of the power and gain.  The ideal is to share a single path
and use another laser frequency which has a significantly different
response to the atoms.  The two acquisition chains produce two integrated
data signals and these are cross multiplied before being formed into
sums and differences.  No "gains" are calculated.  This also means that
the widths of various signals are reduced because no multipliers are needed.
However, I have yet to remove the multiplier signals from the software.