Changed input pullup resistors to pulldown, allows software starts
Input ports had pullup resistors which meant that the software
triggers couldn't start the device as they looked for a rising
edge.  Changed the input resistors to pulldown to fix this problem.

Also, triggers and reset are now high for 2 sysClk cycles to ensure
they are registered by the modules.  This should have no problems
because all triggers are registered by edge detection and not state
detection.